Class Updates
Added a couple of classes. Calculator is the general interface for
handling equations and the like. For now, just pass along an expression
to the calculator. Eventually, it'll break it down into "expressions,"
smaller units of the equation which it can then evaluate. Partial
separation of expressions is functional at the moment, but it doesn't
really do anything. For example, if you type in "3x+5=10x", it should
separate out to "3", "x", "+", "5", "=", "10", and "x". The numbers are
separated, the operators are, and the variables are, too. Very, very
early version just to demonstrate, will [s]likely[/s] [s]certainly[/s]
[s]for sure[/s] absolutely be redone later.