LARGE changes. First version of expression parsing and error checking.
Wrote a significant portion of the parsing of a given expression, which
is correct as far as I can tell. There's some minor modifications that
need to happen to prevent infinite recursion, but since it's not calling
itself for now, it's stable. I also began a text document that explains
what I have planned, which walks through a brief list of the flow of an
expression through the equation. I have currently been testing the
parsing with "5*(3x+1)+5", and it *should* be properly working for a
much wider array of expressions. Feel free to test, and let me know of
any bugs!