Added Current Status and To-Do List
Added a document describing what's been done and what is to be done.
Stuff should be added and edited on here, but don't delete things off if
they're old. Somebody reading it should be able to get a good idea of
where the project's status is at the moment, and what needs to be done
next.