Mostly done parsing, started evaluation
Fixed the parsing problem! Parsing is pretty much done, although I can't
tell quite if it's completely correct. Exponentiation has begun,
although it needs tweaking because exponentiation is weird - if there
are several in a row, it goes "top down," meaning from right-to-left in
operations, instead of the standard straight first-come-first-server the
rest of the operations follow.